refactor(ExhangeApi): mod to class approach with abstract class
